requirements-dev.txt
# Direct dependencies in addition to requirements.txt:
#  black boto3-stubs[ec2] coverage grpcio-tools isort mypy pylint pylint-pydantic pytest pytest-cov python_on_whales types-protobuf types-PyYAML types-requests

-r requirements.txt

astroid==2.15.5
black==23.3.0
boto3==1.26.153
boto3-stubs==1.26.153
botocore==1.29.153
botocore-stubs==1.29.153
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
dill==0.3.6
grpcio==1.54.2
grpcio-tools==1.54.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.3.0
mypy-boto3-ec2==1.26.153
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
pluggy==1.0.0
protobuf==4.23.3
pydantic==1.10.9
pylint==2.17.4
pylint-plugin-utils==0.8.2
pylint-pydantic==0.1.8
pytest==7.3.2
pytest-cov==4.1.0
python-dateutil==2.8.2
python-on-whales==0.61.0
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
tomlkit==0.11.8
tqdm==4.65.0
typer==0.9.0
types-awscrt==0.16.19
types-protobuf==4.23.0.1
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-s3transfer==0.6.1
types-urllib3==1.26.25.13
typing_extensions==4.6.3
urllib3==1.26.16
wrapt==1.15.0
